db schemas: use BIGINT (Long)

+++ distribution/src/database/openfire_db2.sql
@@ -37,7 +37,7 @@ CREATE INDEX ofUserFlag_eTime_idx ON ofUserFlag (endTime ASC);
 
 CREATE TABLE ofOffline (
   username              VARCHAR(64)     NOT NULL,
-  messageID             INTEGER         NOT NULL,
+  messageID             BIGINT          NOT NULL,
   creationDate          CHAR(15)        NOT NULL,
   messageSize           INTEGER         NOT NULL,
   stanza                VARCHAR(2000)   NOT NULL,
@@ -52,7 +52,7 @@ CREATE TABLE ofPresence (
 );
 
 CREATE TABLE ofRoster (
-  rosterID              INTEGER         NOT NULL,
+  rosterID              BIGINT          NOT NULL,
   username              VARCHAR(64)     NOT NULL,
   jid                   VARCHAR(2000)   NOT NULL,
   sub                   INTEGER         NOT NULL,
@@ -67,7 +67,7 @@ CREATE INDEX ofRoster_jid_idx ON ofRoster (jid ASC);
 
 
 CREATE TABLE ofRosterGroups (
-  rosterID              INTEGER         NOT NULL,
+  rosterID              BIGINT          NOT NULL,
   rank                  INTEGER         NOT NULL,
   groupName             VARCHAR(255)    NOT NULL,
   CONSTRAINT ofRosterGroups_pk PRIMARY KEY (rosterID, rank)
@@ -107,13 +107,13 @@ CREATE TABLE ofGroupUser (
 
 CREATE TABLE ofID (
   idType                INTEGER         NOT NULL,
-  id                    INTEGER         NOT NULL,
+  id                    BIGINT          NOT NULL,
   CONSTRAINT ofID_pk PRIMARY KEY (idType)
 );
 
 
 CREATE TABLE ofProperty (
-  name        VARCHAR(100) NOT NULL,
+  name        VARCHAR(100)  NOT NULL,
   propValue   VARCHAR(3000) NOT NULL,
   encrypted   INTEGER,
   iv          CHAR(24),
@@ -155,7 +155,7 @@ CREATE TABLE ofPrivacyList (
 CREATE INDEX ofPrivacyList_default_idx ON ofPrivacyList (username, isDefault);
 
 CREATE TABLE ofSecurityAuditLog (
-  msgID                 INTEGER         NOT NULL,
+  msgID                 BIGINT          NOT NULL,
   username              VARCHAR(64)     NOT NULL,
   entryStamp            BIGINT          NOT NULL,
   summary               VARCHAR(255)    NOT NULL,
@@ -169,7 +169,7 @@ CREATE INDEX ofSecAuditLog_uname_idx ON ofSecurityAuditLog (username);
 -- MUC tables
 
 CREATE TABLE ofMucService (
-  serviceID           INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
   subdomain           VARCHAR(255)  NOT NULL,
   description         VARCHAR(255),
   isHidden            INTEGER       NOT NULL,
@@ -178,15 +178,15 @@ CREATE TABLE ofMucService (
 CREATE INDEX ofMucService_serviceid_idx ON ofMucService(serviceID);
 
 CREATE TABLE ofMucServiceProp (
-  serviceID           INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
   name                VARCHAR(100)  NOT NULL,
   propValue           VARCHAR(2000) NOT NULL,
   CONSTRAINT ofMucSrvProp_pk PRIMARY KEY (serviceID, name)
 );
 
 CREATE TABLE ofMucRoom (
-  serviceID           INTEGER       NOT NULL,
-  roomID              INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
+  roomID              BIGINT        NOT NULL,
   creationDate        CHAR(15)      NOT NULL,
   modificationDate    CHAR(15)      NOT NULL,
   name                VARCHAR(50)   NOT NULL,
@@ -222,14 +222,14 @@ CREATE INDEX ofMucRoom_srvid_idx ON ofMucRoom (serviceID);
 
 
 CREATE TABLE ofMucRoomProp (
-  roomID                INTEGER         NOT NULL,
+  roomID                BIGINT          NOT NULL,
   name                  VARCHAR(100)    NOT NULL,
   propValue             VARCHAR(2000)   NOT NULL,
   CONSTRAINT ofMucRoomProp_pk PRIMARY KEY (roomID, name)
 );
 
 CREATE TABLE ofMucRoomRetiree (
-  serviceID           INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
   name                VARCHAR(50)   NOT NULL,
   alternateJID        VARCHAR(2000),
   reason              VARCHAR(1024),
@@ -238,7 +238,7 @@ CREATE TABLE ofMucRoomRetiree (
 );
 
 CREATE TABLE ofMucAffiliation (
-  roomID              INTEGER       NOT NULL,
+  roomID              BIGINT        NOT NULL,
   jid                 VARCHAR(1000) NOT NULL,
   affiliation         INTEGER       NOT NULL,
   CONSTRAINT ofMucAffil_pk PRIMARY KEY (roomID, jid)
@@ -246,7 +246,7 @@ CREATE TABLE ofMucAffiliation (
 
 
 CREATE TABLE ofMucMember (
-  roomID              INTEGER       NOT NULL,
+  roomID              BIGINT        NOT NULL,
   jid                 VARCHAR(1000) NOT NULL,
   nickname            VARCHAR(255),
   firstName           VARCHAR(100),
@@ -259,8 +259,8 @@ CREATE TABLE ofMucMember (
 
 
 CREATE TABLE ofMucConversationLog (
-  roomID              INTEGER       NOT NULL,
-  messageID         INTEGER NOT NULL,
+  roomID              BIGINT        NOT NULL,
+  messageID           BIGINT        NOT NULL,
   sender              VARCHAR(2000) NOT NULL,
   nickname            VARCHAR(255),
   logTime             CHAR(15)      NOT NULL,
