+++ distribution/src/database/openfire_sqlserver.sql
@@ -37,7 +37,7 @@ CREATE INDEX ofUserFlag_eTime_idx ON ofUserFlag (endTime ASC);
 
 CREATE TABLE ofOffline (
   username              NVARCHAR(64)    NOT NULL,
-  messageID             INTEGER         NOT NULL,
+  messageID             BIGINT          NOT NULL,
   creationDate          CHAR(15)        NOT NULL,
   messageSize           INTEGER         NOT NULL,
   stanza                NTEXT           NOT NULL,
@@ -54,7 +54,7 @@ CREATE TABLE ofPresence (
 
 
 CREATE TABLE ofRoster (
-  rosterID              INTEGER         NOT NULL,
+  rosterID              BIGINT          NOT NULL,
   username              NVARCHAR(64)    NOT NULL,
   jid                   NVARCHAR(1024)  NOT NULL,
   sub                   INTEGER         NOT NULL,
@@ -69,7 +69,7 @@ CREATE INDEX ofRoster_jid_idx ON ofRoster (jid ASC);
 
 
 CREATE TABLE ofRosterGroups (
-  rosterID              INTEGER         NOT NULL,
+  rosterID              BIGINT          NOT NULL,
   rank                  INTEGER         NOT NULL,
   groupName             NVARCHAR(255)   NOT NULL,
   CONSTRAINT ofRosterGroups_pk PRIMARY KEY (rosterID, rank)
@@ -110,7 +110,7 @@ CREATE TABLE ofGroupUser (
 
 CREATE TABLE ofID (
   idType                INTEGER         NOT NULL,
-  id                    INTEGER         NOT NULL,
+  id                    BIGINT          NOT NULL,
   CONSTRAINT ofID_pk PRIMARY KEY (idType)
 );
 
@@ -155,7 +155,7 @@ CREATE TABLE ofPrivacyList (
 CREATE INDEX ofPrivacyList_default_idx ON ofPrivacyList (username, isDefault);
 
 CREATE TABLE ofSecurityAuditLog (
-  msgID                 INTEGER         NOT NULL,
+  msgID                 BIGINT          NOT NULL,
   username              NVARCHAR(64)    NOT NULL,
   entryStamp            BIGINT          NOT NULL,
   summary               NVARCHAR(255)   NOT NULL,
@@ -169,7 +169,7 @@ CREATE INDEX ofSecurityAuditLog_uname_idx ON ofSecurityAuditLog (username);
 -- MUC tables
 
 CREATE TABLE ofMucService (
-  serviceID           INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
   subdomain           NVARCHAR(255) NOT NULL,
   description         NVARCHAR(255),
   isHidden            INTEGER       NOT NULL,
@@ -178,15 +178,15 @@ CREATE TABLE ofMucService (
 CREATE INDEX ofMucService_serviceid_idx ON ofMucService(serviceID);
 
 CREATE TABLE ofMucServiceProp (
-  serviceID           INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
   name                NVARCHAR(100) NOT NULL,
   propValue           NVARCHAR(2000) NOT NULL,
   CONSTRAINT ofMucServiceProp_pk PRIMARY KEY (serviceID, name)
 );
 
 CREATE TABLE ofMucRoom (
-  serviceID           INTEGER       NOT NULL,
-  roomID              INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
+  roomID              BIGINT        NOT NULL,
   creationDate        CHAR(15)      NOT NULL,
   modificationDate    CHAR(15)      NOT NULL,
   name                NVARCHAR(50)  NOT NULL,
@@ -221,14 +221,14 @@ CREATE INDEX ofMucRoom_roomid_idx ON ofMucRoom(roomID);
 CREATE INDEX ofMucRoom_serviceid_idx ON ofMucRoom(serviceID);
 
 CREATE TABLE ofMucRoomProp (
-  roomID                INTEGER         NOT NULL,
+  roomID                BIGINT          NOT NULL,
   name                  NVARCHAR(100)   NOT NULL,
   propValue             NVARCHAR(2000)  NOT NULL,
   CONSTRAINT ofMucRoomProp_pk PRIMARY KEY (roomID, name)
 );
 
 CREATE TABLE ofMucRoomRetiree (
-  serviceID           INTEGER       NOT NULL,
+  serviceID           BIGINT        NOT NULL,
   name                NVARCHAR(50)  NOT NULL,
   alternateJID        NVARCHAR(2000),
   reason              NVARCHAR(1024),
@@ -237,14 +237,14 @@ CREATE TABLE ofMucRoomRetiree (
 );
 
 CREATE TABLE ofMucAffiliation (
-  roomID              INTEGER        NOT NULL,
+  roomID              BIGINT         NOT NULL,
   jid                 NVARCHAR(424) NOT NULL,
   affiliation         INTEGER        NOT NULL,
   CONSTRAINT ofMucAffiliation_pk PRIMARY KEY (roomID,jid)
 );
 
 CREATE TABLE ofMucMember (
-  roomID              INTEGER        NOT NULL,
+  roomID              BIGINT         NOT NULL,
   jid                 NVARCHAR(424)  NOT NULL,
   nickname            NVARCHAR(255),
   firstName           NVARCHAR(100),
@@ -256,8 +256,8 @@ CREATE TABLE ofMucMember (
 );
 
 CREATE TABLE ofMucConversationLog (
-  roomID              INTEGER        NOT NULL,
-  messageID         INTEGER       NOT NULL,
+  roomID              BIGINT         NOT NULL,
+  messageID           BIGINT         NOT NULL,
   sender              NVARCHAR(1024) NOT NULL,
   nickname            NVARCHAR(255),
   logTime             CHAR(15)       NOT NULL,
